src/server/mcp_hub/gdocs/requirements.txt
fastmcp
python-dotenv
motor
cryptography
google-api-python-client
google-auth-oauthlib
